# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      summary: List consultations
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
        - name: doctorId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Consultations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationListResponse"
    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Start a consultation record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationInput"
      responses:
        "201":
          description: Consultation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}:
    get:
      operationId: getConsultation
      tags: [consultations]
      summary: Get full consultation details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
    patch:
      operationId: updateConsultation
      tags: [consultations]
      summary: Update consultation notes and diagnosis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationUpdate"
      responses:
        "200":
          description: Updated consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}/complete:
    post:
      operationId: completeConsultation
      tags: [consultations]
      summary: Mark consultation as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteConsultationInput"
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"

  # ── PRESCRIPTIONS ─────────────────────────────────────────────────────────
...
  /dashboard/revenue-chart:
    get:
      operationId: getRevenueChart
      tags: [dashboard]
      summary: Revenue by day for the last N days
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Revenue chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueChartPoint"
  /dashboard/patient-trends:
    get:
      operationId: getPatientTrends
      tags: [dashboard]
      summary: Patient visit trends by day/week
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Patient trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"

  # ── REPORTS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Consultation:
      type: object
      required: [id, patientId, doctorId, status, visitDate, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        patientName:
          type: string
        doctorId:
          type: string
        doctorName:
          type: string
        appointmentId:
          type: ["string", "null"]
        tokenId:
          type: ["string", "null"]
        visitDate:
          type: string
        status:
          type: string
          enum: [in_progress, completed, cancelled]
        chiefComplaint:
          type: ["string", "null"]
        historyOfPresentIllness:
          type: ["string", "null"]
        clinicalNotes:
          type: ["string", "null"]
        soapSubjective:
          type: ["string", "null"]
        soapObjective:
          type: ["string", "null"]
        soapAssessment:
          type: ["string", "null"]
        soapPlan:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        icd10Code:
          type: ["string", "null"]
        advice:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        followUpNotes:
          type: ["string", "null"]
        investigationOrders:
          type: ["string", "null"]
        vitals:
          type: ["object", "null"]
          additionalProperties: true
        createdAt:
          type: string
        updatedAt:
          type: string
    ConsultationInput:
      type: object
      required: [patientId, doctorId]
      properties:
        patientId:
          type: string
        doctorId:
          type: string
        appointmentId:
          type: string
        tokenId:
          type: string
        chiefComplaint:
          type: string
    ConsultationUpdate:
      type: object
      properties:
        chiefComplaint:
          type: string
        historyOfPresentIllness:
          type: string
        clinicalNotes:
          type: string
        soapSubjective:
          type: string
        soapObjective:
          type: string
        soapAssessment:
          type: string
        soapPlan:
          type: string
        diagnosis:
          type: string
        icd10Code:
          type: string
        advice:
          type: string
        followUpDate:
          type: string
        followUpNotes:
          type: string
        investigationOrders:
          type: string
    CompleteConsultationInput:
      type: object
      properties:
        diagnosis:
          type: string
        advice:
          type: string
        followUpDate:
          type: string
    ConsultationListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Consultation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── PRESCRIPTION SCHEMAS ─────────────────────────────────────────────────
...
    RevenueChartPoint:
      type: object
      required: [date, revenue, count]
      properties:
        date:
          type: string
        revenue:
          type: number
        count:
          type: integer
    TrendPoint:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer
        newPatients:
          type: integer

    # ── REPORT SCHEMAS ───────────────────────────────────────────────────────
...